@@ -1,1 +1,1 @@
-ruby-1.9.3-p125
+ruby-2.0.0-p0
