@@ -23,15 +23,15 @@ class UsersController < ApplicationController
     sort_val = get_sort
     query = params[:uq].blank? ? nil : "%"+params[:uq].downcase+"%"
     @users = User.where("(login like lower(?) OR display_name like lower(?)) and ldap_available = true and login not in ('testppp','tstem31')",
-     query,query).paginate(:page => params[:page], :per_page => 10, :order => sort_val) unless query.blank?   
+                        query,query).paginate(:page => params[:page], :per_page => 10, :order => sort_val) unless query.blank?
     @users = User.where("ldap_available = true and login not in ('testapp','tstem31')").paginate(:page => params[:page], :per_page => 10, :order => sort_val) if query.blank?
   end
 
   # Display user profile
   def show
     if @user.respond_to? :profile_events
-      @events = @user.profile_events(100) 
-    else 
+      @events = @user.profile_events(100)
+    else
       @events = []
     end
 
@@ -39,13 +39,12 @@ class UsersController < ApplicationController
     begin
       problem_index = 0
       num_retry += 1
-      @trophies=[]
-      @user.trophies.each do |t| 
+      @trophies = []
+      @user.trophies.each do |t|
         problem_index += 1
         @trophies << GenericFile.find("scholarsphere:#{t.generic_file_id}")
       end
       rescue ActiveFedora::ObjectNotFoundError => e
-        puts ">>>>>>>>>>>>>>>>>>>>>>>>>>>>show" + e.inspect
         loop_counter = 0
         @user.trophies.each do |t|
           loop_counter += 1
@@ -58,7 +57,7 @@ class UsersController < ApplicationController
         else
           raise
         end
-    end       
+    end
     @followers = @user.followers
     @following = @user.all_following
     @linkedInUrl = @user.linkedin_handle
@@ -70,19 +69,18 @@ class UsersController < ApplicationController
   def edit
     @user = current_user
     @groups = @user.groups
-    @trophies=[]
+    @trophies = []
 
     num_retry = 0
     begin
       problem_index = 0
       num_retry += 1
-      @trophies=[]
-      @user.trophies.each do |t| 
+      @trophies = []
+      @user.trophies.each do |t|
         problem_index += 1
         @trophies << GenericFile.find("scholarsphere:#{t.generic_file_id}")
       end
       rescue ActiveFedora::ObjectNotFoundError => e
-        puts ">>>>>>>>>>>>>>>>>>>>>>>>>>>>edit" + e.inspect
         loop_counter = 0
         @user.trophies.each do |t|
           loop_counter += 1
@@ -95,44 +93,46 @@ class UsersController < ApplicationController
         else
           raise
         end
-    end    
+    end
   end
 
   # Process changes from profile form
   def update
     @user.update_attributes(params[:user])
-  
     @user.populate_attributes if params[:update_directory]
     @user.avatar = nil if params[:delete_avatar]
     unless @user.save
       redirect_to edit_profile_path(@user.to_s), alert: @user.errors.full_messages
       return
     end
-    delete_trophy = params.keys.reject{|k,v|k.slice(0,'remove_trophy'.length)!='remove_trophy'}
-    delete_trophy = delete_trophy.map{|v| v.slice('remove_trophy_'.length..-1)}
-    delete_trophy.each do | smash_trophy |
+    delete_trophy = params.keys.reject{ |k,v| k.slice(0,'remove_trophy'.length)!='remove_trophy' }
+    delete_trophy = delete_trophy.map{ |v| v.slice('remove_trophy_'.length..-1) }
+    delete_trophy.each do |smash_trophy|
       Trophy.where(user_id: current_user.id, generic_file_id: smash_trophy.slice('scholarsphere:'.length..-1)).each.map(&:delete)
     end
     Sufia.queue.push(UserEditProfileEventJob.new(@user.login))
     redirect_to profile_path(@user.to_s), notice: "Your profile has been updated"
   end
-  def toggle_trophy    
-     unless current_user.can? :edit, permissions_solr_doc_for_id("scholarsphere:#{params[:file_id]}")
-       return false
-     end
-     # TO DO  make sure current user has access to file
-     t = Trophy.where(:generic_file_id => params[:file_id], :user_id => current_user.id).first
-     if t.blank? 
-       t = Trophy.create(:generic_file_id => params[:file_id], :user_id => current_user.id)
-       return false unless t.persisted?
-     else
-       t.delete  
-       #TODO do this better says Mike
-       return false if t.persisted?  
-     end
-     render :json => t
-  end 
 
+  def toggle_trophy
+     id = params[:file_id]
+     id = "#{Sufia::Engine.config.id_namespace}:#{id}" unless id.include?(":")
+    unless current_user.can? :edit, id
+      redirect_to root_path, alert: "You do not have permissions to the file"
+      return false
+    end
+    # TODO:  make sure current user has access to file
+    t = Trophy.where(:generic_file_id => params[:file_id], :user_id => current_user.id).first
+    if t.blank?
+      t = Trophy.create(:generic_file_id => params[:file_id], :user_id => current_user.id)
+      return false unless t.persisted?
+    else
+      t.delete
+      # TODO: do this better says Mike
+      return false if t.persisted?
+    end
+    render :json => t
+  end
 
   # Follow a user
   def follow
@@ -169,10 +169,10 @@ class UsersController < ApplicationController
   def get_sort
     sort = params[:sort].blank? ? "name" : params[:sort]
     sort_val = case sort
-           when "name"  then "display_name"
-           when "name desc"   then "display_name DESC"
-           else sort
-           end
+               when "name"  then "display_name"
+               when "name desc"   then "display_name DESC"
+               else sort
+               end
     return sort_val
   end
 end
